+++ .github/files/jetpack-staging-sites/k6-backend.js
@@ -34,26 +34,26 @@ export default function () {
 			// Jetpack connection test.
 			let res = http.get( `${ site.url }/wp-json/jetpack/v4/connection/test`, params );
 			check( res, {
-				'status was 200': r => r.status == 200,
+				'connection status was 200': r => r.status == 200,
 				'verify connection message': r => r.body.includes( 'All connection tests passed' ),
 			} );
 
 			// Jetpack sync status.
 			res = http.get( `${ site.url }/wp-json/jetpack/v4/sync/status`, params );
 			check( res, {
-				'status was 200': r => r.status == 200,
+				'sync status was 200': r => r.status == 200,
 			} );
 
 			// /wp-json/wp/v2/posts
 			res = http.get( `${ site.url }/wp-json/wp/v2/posts`, params );
 			check( res, {
-				'status was 200': r => r.status == 200,
+				'posts status was 200': r => r.status == 200,
 			} );
 
 			// /wp-json/wp/v2/categories
 			res = http.get( `${ site.url }/wp-json/wp/v2/categories`, params );
 			check( res, {
-				'status was 200': r => r.status == 200,
+				'categories status was 200': r => r.status == 200,
 			} );
 
 			// /wp-json/wp-site-health/v1/tests/dotorg-communication
@@ -62,7 +62,7 @@ export default function () {
 				params
 			);
 			check( res, {
-				'status was 200': r => r.status == 200,
+				'dotorg-communication status was 200': r => r.status == 200,
 				'verify communication message': r =>
 					r.body.includes( 'Can communicate with WordPress.org' ),
 			} );