+++ projects/js-packages/image-guide/src/MeasurableImage.ts
@@ -2,7 +2,7 @@ export type SourceCallbackFn = ( node: HTMLElement ) => string | null;
 export type Dimensions = { width: number; height: number };
 export type Weight = { weight: number };
 
-type FetchFn = ( input: URL | RequestInfo, init?: RequestInit ) => Promise< Response >;
+export type FetchFn = ( input: URL | RequestInfo, init?: RequestInit ) => Promise< Response >;
 
 /**
  * A class that represents a DOM Element that
@@ -13,7 +13,7 @@ export class MeasurableImage {
 	readonly node: HTMLElement | HTMLImageElement;
 	private getURLCallback: SourceCallbackFn;
 
-	public fetch = window.fetch;
+	public fetch: FetchFn;
 
 	/**
 	 * Constructor.
@@ -25,11 +25,11 @@ export class MeasurableImage {
 	constructor(
 		node: HTMLElement | HTMLImageElement,
 		getURL: SourceCallbackFn,
-		fetchFn: FetchFn = fetch
+		fetchFn: FetchFn | null = null
 	) {
 		this.node = node;
 		this.getURLCallback = getURL;
-		this.fetch = fetchFn;
+		this.fetch = fetchFn ?? window.fetch.bind( window );
 	}
 
 	public getURL() {
@@ -94,7 +94,7 @@ export class MeasurableImage {
 	 */
 	private async fetchFileWeight( url: string ) {
 		const response = await this.fetch( url );
-		if ( ! response.url ) {
+		if ( ! response.ok ) {
 			// eslint-disable-next-line no-console
 			console.log( `Can't get image size for ${ url } likely due to a CORS error.` );
 			return -1;
@@ -112,13 +112,29 @@ export class MeasurableImage {
 	private async fetchFileDimensions( url: string ) {
 		const img = new Image();
 		img.src = url;
-		return new Promise< { width: number; height: number } >( resolve => {
+		return new Promise< { width: number; height: number } >( ( resolve, reject ) => {
 			img.onload = () => {
 				resolve( { width: Math.round( img.width ), height: Math.round( img.height ) } );
 			};
 			img.onerror = () => {
-				resolve( { width: -1, height: -1 } );
+				reject( 'Unable to load image.' );
 			};
 		} );
+	}
+
+	/**
+	 * Checks if the image is too small and should be ignored. Will return false on images
+	 * that don't load at all - we can't establish they're tiny!
+	 *
+	 * @returns {boolean} - if the image is smaller than 65 pixels width and height return true
+	 */
+	public async isImageTiny(): Promise< boolean > {
+		try {
+			const minSize = 65;
+			const dimensions = await this.fetchFileDimensions( this.getURL() );
+			return dimensions.width < minSize || dimensions.height < minSize;
+		} catch ( err ) {
+			return false;
+		}
 	}
 }