+++ projects/packages/my-jetpack/_inc/state/resolvers.js
@@ -12,6 +12,7 @@ import {
 	REST_API_SITE_PURCHASES_ENDPOINT,
 	REST_API_SITE_PRODUCTS_ENDPOINT,
 	REST_API_CHAT_AVAILABILITY_ENDPOINT,
+	REST_API_CHAT_AUTHENTICATION_ENDPOINT,
 	PRODUCTS_THAT_NEEDS_INITIAL_FETCH,
 } from './constants';
 import resolveProductStatsRequest from './stats-resolvers';
@@ -84,6 +85,21 @@ const myJetpackResolvers = {
 				dispatch.setChatAvailabilityIsFetching( false );
 			} catch ( error ) {
 				dispatch.setChatAvailabilityIsFetching( false );
+			}
+		},
+
+	getChatAuthentication:
+		() =>
+		async ( { dispatch } ) => {
+			dispatch.setChatAuthenticationIsFetching( true );
+
+			try {
+				dispatch.setChatAuthentication(
+					await apiFetch( { path: REST_API_CHAT_AUTHENTICATION_ENDPOINT } )
+				);
+				dispatch.setChatAuthenticationIsFetching( false );
+			} catch ( error ) {
+				dispatch.setChatAuthenticationIsFetching( false );
 			}
 		},