+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/RecommendationsMeta.svelte
@@ -3,15 +3,22 @@
 	import { __, sprintf } from '@wordpress/i18n';
 	import Button from '../../elements/Button.svelte';
 	import ErrorNotice from '../../elements/ErrorNotice.svelte';
+	import NoticeIcon from '../../svg/notice-outline.svg';
 	import RefreshIcon from '../../svg/refresh.svg';
 	import MultiProgress from './MultiProgress.svelte';
-	import { requestImageAnalysis, initializeIsaSummary, isaSummary } from './store/isa-summary';
+	import { resetIsaQuery } from './store/isa-data';
+	import {
+		requestImageAnalysis,
+		initializeIsaSummary,
+		isaSummary,
+		ISAStatus,
+	} from './store/isa-summary';
 
 	onMount( () => {
 		initializeIsaSummary();
 	} );
 
-	let errorMessage: undefined | string;
+	let submitError: undefined | string;
 	let requestingReport = false;
 
 	/**
@@ -23,6 +30,26 @@
 	);
 
 	/**
+	 * Work out if there is an error to show in the UI.
+	 */
+	$: errorMessage =
+		submitError ||
+		( $isaSummary?.status === ISAStatus.Stuck &&
+			__(
+				'Your Image Size Analysis task seems to have gotten stuck, or our system is under unusual load. Please try again. If the issue persists, please contact support.',
+				'jetpack-boost'
+			) );
+
+	/**
+	 * Work out whether we have a 'give us a minute' notice to show.
+	 */
+	$: waitNotice =
+		( requestingReport && __( 'Getting ready…', 'jetpack-boost' ) ) ||
+		( $isaSummary?.status === ISAStatus.New && __( 'Warming up the engine…', 'jetpack-boost' ) ) ||
+		( $isaSummary?.status === ISAStatus.Queued &&
+			__( 'Give us a few minutes while we go through your images…', 'jetpack-boost' ) );
+
+	/**
 	 * Start a new image analysis job.
 	 */
 	async function onStartAnalysis() {
@@ -30,6 +57,7 @@
 			errorMessage = undefined;
 			requestingReport = true;
 			await requestImageAnalysis();
+			resetIsaQuery();
 		} catch ( err ) {
 			errorMessage = err.message;
 		} finally {
@@ -42,32 +70,26 @@
 	<div class="summary">
 		{__( 'Loading…', 'jetpack-boost' )}
 	</div>
-{:else if $isaSummary.status === 'not-found'}
-	<div class="button-area">
-		<Button disabled={requestingReport} on:click={onStartAnalysis}>Start image analysis</Button>
-	</div>
-{:else if $isaSummary.status === 'new'}
-	<!-- Boost fetches a list of the site's content between creating the report and enqueuing the job. Status = new. -->
-	<div class="wait-notice">
-		{__( 'Give us a few minutes while we go through your content…', 'jetpack-boost' )}
-	</div>
-{:else if $isaSummary.status === 'error'}
-	<!-- @TODO -->
-	<pre>
-		{JSON.stringify( $isaSummary )}
-	</pre>
 {:else}
-	<!-- Boost is either generating a report, or has finished one. -->
-	{@const  inProgress = $isaSummary.status === 'queued' }
-
-	{#if inProgress}
-		<div class="summary-line wait-notice">
-			{__( 'Give us a few minutes while we go through your images…', 'jetpack-boost' )}
+	<!-- Show error messages or "please wait" messages. -->
+	{#if errorMessage}
+		<div class="error-area">
+			<ErrorNotice title={__( 'Something has gone wrong.', 'jetpack-boost' )}>
+				{errorMessage}
+			</ErrorNotice>
 		</div>
-	{:else}
+	{:else if waitNotice}
+		<div class="summary-line wait-notice">
+			{waitNotice}
+		</div>
+	{/if}
+
+	<!-- Show a summary line if the report is completed. -->
+	{#if ! requestingReport && $isaSummary.status === ISAStatus.Completed}
 		<div class="summary-line">
 			{#if totalIssues > 0}
 				<div class="has-issues summary">
+					<NoticeIcon class="icon" />
 					{sprintf(
 						/* translators: %d is the number of issues that were found */
 						__( 'Found a total of %d issues', 'jetpack-boost' ),
@@ -92,28 +114,37 @@
 		</div>
 	{/if}
 
-	{#if ! requestingReport && ! errorMessage}
+	<!-- Show progress if a job is rolling. -->
+	{#if ! requestingReport && [ ISAStatus.Completed, ISAStatus.Queued ].includes( $isaSummary.status )}
 		<MultiProgress />
+	{/if}
 
+	<!-- Show a button to view the report if it's in progress or completed. -->
+	{#if [ ISAStatus.Queued, ISAStatus.Completed ].includes( $isaSummary.status ) && ! requestingReport}
 		<div class="button-area">
 			<Button href="#image-size-analysis/all/1" disabled={requestingReport}>
-				{inProgress
-					? __( 'See report in progress', 'jetpack-boost' )
-					: __( 'See full report', 'jetpack-boost' )}
+				{$isaSummary.status === ISAStatus.Completed
+					? __( 'See full report', 'jetpack-boost' )
+					: __( 'View report in progress', 'jetpack-boost' )}
+			</Button>
+		</div>
+	{/if}
+
+	<!-- Show a button to kick off a report -->
+	{#if ! [ ISAStatus.New, ISAStatus.Queued, ISAStatus.Completed ].includes( $isaSummary.status )}
+		<div class="button-area">
+			<Button disabled={requestingReport} on:click={onStartAnalysis}>
+				{$isaSummary.status === ISAStatus.Completed
+					? __( 'Analyze again', 'jetpack-boost' )
+					: __( 'Start image analysis', 'jetpack-boost' )}
 			</Button>
 		</div>
 	{/if}
 {/if}
 
-{#if errorMessage}
-	<div class="error-area">
-		<ErrorNotice title={__( 'Failed to request an Image Analysis job', 'jetpack-boost' )}>
-			{errorMessage}
-		</ErrorNotice>
-	</div>
-{/if}
+<style lang="scss">
+	@use '../../../css/main/variables.scss' as *;
 
-<style>
 	.summary-line {
 		font-size: 14px;
 		line-height: 22px;
@@ -121,16 +152,39 @@
 		flex-direction: row;
 		align-items: flex-start;
 		margin-bottom: 17px;
+
+		@media ( max-width: 600px ) {
+			flex-direction: column;
+		}
+	}
+
+	.summary-line button {
+		:global( svg ) {
+			margin: 4px 4px 2px 0;
+			fill: $jetpack-green;
+		}
+
+		@media ( max-width: 600px ) {
+			margin-top: 15px;
+		}
 	}
 
 	.summary {
 		margin-right: 5px;
 		flex-grow: 1;
 		position: relative;
+		color: $jetpack-green;
 	}
 
 	.has-issues {
-		color: var( --red_50 );
+		color: $red_50;
+	}
+
+	.has-issues :global( svg ) {
+		width: 22px;
+		height: 22px;
+		top: 4px;
+		position: relative;
 	}
 
 	.wait-notice {