+++ projects/plugins/boost/app/assets/src/js/pages/getting-started/GettingStarted.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-	import { onMount } from 'svelte';
-	import { derived, get, writable } from 'svelte/store';
+	import { derived, writable } from 'svelte/store';
 	import { Snackbar } from '@wordpress/components';
 	import ActivateLicense from '../../elements/ActivateLicense.svelte';
 	import ReactComponent from '../../elements/ReactComponent.svelte';
@@ -8,7 +7,6 @@
 	import Header from '../../sections/Header.svelte';
 	import config, { markGetStartedComplete } from '../../stores/config';
 	import { connection } from '../../stores/connection';
-	import { modulesState, modulesStatePending } from '../../stores/modules';
 	import { recordBoostEvent } from '../../utils/analytics';
 	import { getUpgradeURL } from '../../utils/upgrade';
 
@@ -32,90 +30,72 @@
 		}
 	);
 
-	const ensureConnection = async () => {
-		let connectionStore;
-		connection.subscribe( value => {
-			connectionStore = value;
-		} );
+	/**
+	 * Mark that getting started is completed, and head to the next page.
+	 *
+	 * @param {string} externalPage If specified, head to the external page instead of staying on the dashboard.
+	 */
+	function finishGettingStarted( externalPage?: string ) {
+		markGetStartedComplete();
 
-		if ( connectionStore.connected ) {
-			return;
-		}
-
-		await connection.initialize();
-
-		if ( ! connectionStore.connected ) {
-			throw connectionStore.error;
-		}
-	};
-
-	const chooseFreePlan = async () => {
-		initiatingFreePlan = true;
-
-		await Promise.all( [
-			recordBoostEvent( 'free_cta_from_getting_started_page_in_plugin', {} ),
-			await ( async () => {
-				try {
-					await ensureConnection();
-
-					// Allow opening the boost settings page. The actual flag is changed in the backend by enabling the critical-css module below.
-					markGetStartedComplete();
-
-					// Wait for the module to become active.
-					// Otherwise Critical CSS Generation will fail to start.
-					$modulesState.critical_css.active = true;
-					const unsubscribe = modulesStatePending.subscribe( isPending => {
-						if ( isPending === false ) {
-							unsubscribe();
-							navigate( '/' );
-						}
-					} );
-				} catch ( e ) {
-					dismissedSnackbar.set( false );
-				} finally {
-					initiatingFreePlan = false;
-				}
-			} )(),
-		] );
-	};
-
-	const choosePaidPlan = async () => {
-		initiatingPaidPlan = true;
-
-		await Promise.all( [
-			await recordBoostEvent( 'premium_cta_from_getting_started_page_in_plugin', {} ),
-			await ( async () => {
-				try {
-					await ensureConnection();
-
-					// Check if the site is already on a premium plan and go directly to settings if so.
-					if ( get( config ).isPremium ) {
-						// Allow opening the boost settings page.
-						markGetStartedComplete();
-
-						navigate( '/', { replace: true } );
-						return;
-					}
-
-					window.location.href = getUpgradeURL();
-				} catch ( e ) {
-					dismissedSnackbar.set( false );
-				} finally {
-					initiatingPaidPlan = false;
-				}
-			} )(),
-		] );
-	};
-
-	onMount( () => {
-		// If we don't have pricing data, we should skip the page and go directly to settings.
-		if ( typeof pricing.yearly === 'undefined' ) {
-			// Allow opening the boost settings page.
-			markGetStartedComplete();
-
+		if ( externalPage ) {
+			window.location.href = externalPage;
+		} else {
 			navigate( '/', { replace: true } );
 		}
-	} );
+	}
+
+	/**
+	 * User clicked "Free plan"
+	 */
+	async function chooseFreePlan() {
+		initiatingFreePlan = true;
+
+		try {
+			// Make sure there is a Jetpack connection and record this selection.
+			await Promise.all( [
+				connection.initialize(),
+				recordBoostEvent( 'free_cta_from_getting_started_page_in_plugin', {} ),
+			] );
+
+			// Head to the settings page.
+			finishGettingStarted();
+		} catch ( e ) {
+			// Un-dismiss snackbar on error. Actual error comes from connection object.
+			dismissedSnackbar.set( false );
+		} finally {
+			initiatingFreePlan = false;
+		}
+	}
+
+	/**
+	 * User clicked Premium.
+	 */
+	async function choosePaidPlan() {
+		initiatingPaidPlan = true;
+
+		try {
+			// Make sure there is a Jetpack connection and record this selection.
+			await Promise.all( [
+				connection.initialize(),
+				recordBoostEvent( 'premium_cta_from_getting_started_page_in_plugin', {} ),
+			] );
+
+			// Check if the site is already on a premium plan and go directly to settings if so.
+			if ( $config.isPremium ) {
+				finishGettingStarted();
+				return;
+			}
+
+			// Go to the purchase flow.
+			finishGettingStarted( getUpgradeURL() );
+		} catch ( e ) {
+			// Un-dismiss snackbar on error. Actual error comes from connection object.
+			dismissedSnackbar.set( false );
+		} finally {
+			initiatingPaidPlan = false;
+		}
+	}
 </script>
 
 <div id="jb-dashboard" class="jb-dashboard jb-dashboard--main">
@@ -123,29 +103,27 @@
 		<ActivateLicense />
 	</Header>
 
-	{#if pricing.yearly}
-		<div class="jb-section jb-section--alt">
-			<div class="jb-container">
-				<div class="jb-pricing-table">
+	<div class="jb-section jb-section--alt">
+		<div class="jb-container">
+			<div class="jb-pricing-table">
+				<ReactComponent
+					this={BoostPricingTable}
+					{pricing}
+					onPremiumCTA={choosePaidPlan}
+					onFreeCTA={chooseFreePlan}
+					chosenFreePlan={initiatingFreePlan}
+					chosenPaidPlan={initiatingPaidPlan}
+				/>
+				{#if $snackbarMessage}
 					<ReactComponent
-						this={BoostPricingTable}
-						{pricing}
-						onPremiumCTA={choosePaidPlan}
-						onFreeCTA={chooseFreePlan}
-						chosenFreePlan={initiatingFreePlan}
-						chosenPaidPlan={initiatingPaidPlan}
+						this={Snackbar}
+						children={$snackbarMessage}
+						onDismiss={() => dismissedSnackbar.set( true )}
 					/>
-					{#if $snackbarMessage}
-						<ReactComponent
-							this={Snackbar}
-							children={$snackbarMessage}
-							onDismiss={() => dismissedSnackbar.set( true )}
-						/>
-					{/if}
-				</div>
+				{/if}
 			</div>
 		</div>
-	{/if}
+	</div>
 </div>
 
 <style lang="scss">