+++ projects/plugins/jetpack/modules/sharedaddy/admin-sharing.js
@@ -166,7 +166,9 @@
 							! $( this ).hasClass( 'preview-press-this' ) &&
 							! $( this ).hasClass( 'preview-email' ) &&
 							! $( this ).hasClass( 'preview-mastodon' ) &&
+							! $( this ).hasClass( 'preview-nextdoor' ) &&
 							! $( this ).hasClass( 'preview-print' ) &&
+							! $( this ).hasClass( 'preview-reddit' ) &&
 							! $( this ).hasClass( 'preview-telegram' ) &&
 							! $( this ).hasClass( 'preview-jetpack-whatsapp' ) &&
 							! $( this ).hasClass( 'share-custom' ) &&