+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,35 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.40.2] - 2023-07-11
+### Changed
+- Updated package dependencies. [#31785]
+
+## [0.40.1] - 2023-07-05
+### Changed
+- Adjust component pricing slider border and box-shadow styling. [#31593]
+- Updated package dependencies. [#31659]
+- Updated package dependencies. [#31661]
+- Update storybook mdx to use `@storybook/blocks` directly rather than `@storybook/addon-docs`. [#31607]
+
+### Fixed
+- Add tagName prop to Col and Container [#31606]
+- Fix Jetpack footer accessibility issues [#31417]
+
+## [0.40.0] - 2023-06-26
+### Added
+- Add authentication to Zendesk chat. [#31339]
+
+## [0.39.0] - 2023-06-23
+### Added
+- Add component Pricing Slider for Stat pricing page.
+- Added config values to publish mirror repo
+- Auto-publish jetpack-components package to npm
+
+## 0.38.1 - 2023-06-21
+### Changed
+- Updated package dependencies. [#31468]
+
 ## 0.38.0 - 2023-06-15
 ### Added
 - Add testimonial component and use it on the backup connect screen [#31221]
@@ -733,3 +762,8 @@
 
 ### Changed
 - Update node version requirement to 14.16.1
+
+[0.40.2]: https://github.com/Automattic/jetpack-components/compare/0.40.1...0.40.2
+[0.40.1]: https://github.com/Automattic/jetpack-components/compare/0.40.0...0.40.1
+[0.40.0]: https://github.com/Automattic/jetpack-components/compare/0.39.0...0.40.0
+[0.39.0]: https://github.com/Automattic/jetpack-components/compare/0.38.1...0.39.0