+++ projects/packages/changelogger/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.6] - 2023-07-10
+- Minor internal updates.
+
+## [3.3.5] - 2023-06-26
+
 ## [3.3.4] - 2023-05-22
 ### Added
 - Set keywords to have `composer require` prompt for `--dev` on installation. [#30756]
@@ -153,6 +158,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - Initial version.
 
+[3.3.6]: https://github.com/Automattic/jetpack-changelogger/compare/3.3.5...3.3.6
+[3.3.5]: https://github.com/Automattic/jetpack-changelogger/compare/3.3.4...3.3.5
 [3.3.4]: https://github.com/Automattic/jetpack-changelogger/compare/3.3.3...3.3.4
 [3.3.3]: https://github.com/Automattic/jetpack-changelogger/compare/3.3.2...3.3.3
 [3.3.2]: https://github.com/Automattic/jetpack-changelogger/compare/3.3.1...3.3.2